+++ src/WebAPI/PaymentRequiredAPI/Controllers/WeatherForecastController.cs
@@ -1,4 +1,4 @@
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
 
 namespace PaymentRequiredAPI.Controllers
 {
@@ -6,7 +6,7 @@ namespace PaymentRequiredAPI.Controllers
     [Route("[controller]")]
     public class WeatherForecastController : ControllerBase
     {
-        private static readonly string[] Summaries = new[]
+        private static readonly string[] s_summaries = new[]
         {
         "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
     };
@@ -25,7 +25,7 @@ namespace PaymentRequiredAPI.Controllers
             {
                 Date = DateTime.Now.AddDays(index),
                 TemperatureC = Random.Shared.Next(-20, 55),
-                Summary = Summaries[Random.Shared.Next(Summaries.Length)]
+                Summary = s_summaries[Random.Shared.Next(s_summaries.Length)]
             })
             .ToArray();
         }