+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [1.1.0] - 2023-07-19
+### Fixed 
+- Change 'NoPaymentRequiredValidator' visibility
+
+### Added
+- 'DemoPaymentRequiredValidator' and 'DemoPaymentRequiredResponseNullObject' for demo purposes.
 
 ## [1.0.1] - 2022-07-01
 ### Fix